requirements.txt
torch
torchvision
torchaudio
torchtext
tensorboard
pyyaml
tqdm
numpy
matplotlib
scikit-learn
datasets
pillow